% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummarizer.R
\name{dataSummarizer}
\alias{dataSummarizer}
\title{Calculate the sum, mean, median and standard deviation of data by group}
\usage{
dataSummarizer(data, groupBy = "CountryName")
}
\arguments{
\item{data}{Object of class "data.frame".}

\item{groupBy}{String specifying what columns to group by.}
}
\value{
An object of class "data.frame".
}
\description{
Calculate the sum, mean, median and standard deviation of data by group
}
\examples{
{
data("measlesdata")
dataSummarizer(measlesdata, groupBy = c("CountryName", "year"))
}
}
