% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastaImporter.R
\name{fastaImporter}
\alias{fastaImporter}
\title{Title}
\usage{
fastaImporter(GenBankIDs, database = "nucleotide")
}
\arguments{
\item{GenBankIDs}{A text file containing the GenBank identifiers of the species of interest, separated by newlines. Comments starting with # allowed.}

\item{database}{String with the name of the database to use.}
}
\value{
An object of class "DNAbin".
}
\description{
Title
}
\examples{
\dontrun{
dna <- fastaImporter(here::here("inst/extdata/GenBankIDs2.txt"))
}
}
