% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotermPlot.R
\name{gotermPlot}
\alias{gotermPlot}
\title{Generate a figure containing the most abundant GO-terms}
\usage{
gotermPlot(
  goterm_results,
  padj_method = "BH",
  filter_from = 5,
  filter_to = 500,
  topamount = 10,
  plot_title
)
}
\arguments{
\item{goterm_results}{Object of class data.frame, output from function gotermAnalysis().}

\item{padj_method}{String defining what adjustment method to use. Options: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{filter_from}{Numeric value that determines the minimum number of times a GO-term needs to appear.}

\item{filter_to}{Numeric value that determines the maximum number of times a GO-term can appear.}

\item{topamount}{Numeric value that determines the number of GO-terms plotted.}

\item{plot_title}{String defining the plot title.}
}
\value{
A object of class "ggplot".
}
\description{
Generate a figure containing the most abundant GO-terms
}
\examples{
\dontrun{
gotermPlot_results <- gotermPlot(gotermAnalysis_results,
topamount = 5,
plot_title = "Top 10 upregulated GO-terms")
}
}
