% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illnessPlot.R
\name{illnessPlot}
\alias{illnessPlot}
\title{Generate a graph showing illness numbers}
\usage{
illnessPlot(data, countries = "Spain", years = 2020, parameter = cases)
}
\arguments{
\item{data}{Object of class "data.frame".}

\item{countries}{String specifying what countries to filter by.}

\item{years}{Integer specifying what years to filter by.}

\item{parameter}{Name of the column of interest, NOT in quotes.}
}
\value{
An object of class "ggplot".
}
\description{
Generate a graph showing illness numbers
}
\examples{
\dontrun{
illnessPlot(data, countries = c("Netherlands", "Belgium", "France"),
years = 2020:2022, parameter = cases)
}
}
