% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotermAnalysis.R
\name{gotermAnalysis}
\alias{gotermAnalysis}
\title{Perform a GO-term enrichment analysis}
\usage{
gotermAnalysis(
  dge_results,
  L2FC = 1,
  padjusted = 0.01,
  ontologytype = "BP",
  pcutoff = 1,
  test_direction = "over",
  upregulated = TRUE
)
}
\arguments{
\item{dge_results}{Object of class "DESeqResults".}

\item{L2FC}{Numeric value of the Log2FoldChange to determine what genes are differently regulated.}

\item{padjusted}{Numeric value of the adjusted p-value to determine what genes are differently regulated.}

\item{ontologytype}{Type of ontology to search for. Options: "BP", "CC", "MF".}

\item{pcutoff}{Numeric value of what p-values to filter by.}

\item{test_direction}{Character string that determines whether to look for over or under represented GO-terms.}

\item{upregulated}{Boolean stating whether to look for the upregulated or downregulated genes.}
}
\value{
An object of class "data.frame".
}
\description{
Perform a GO-term enrichment analysis
}
\examples{
\dontrun{
results_dge <- DESeq2::results(DESeq2::DESeq(DESeqDataSet))
gotermAnalysis(results_dge, ontologytype = "BP", upregulated = FALSE)
}
}
